fix: bind migration reader to verified bytes
